Refactored SubjectAccessRequest package to use one set of Entities rather than an unsynced and alldata. The same with the Repositories.
Removed the application-sar.yml and set helm values to expected env settings
