+++ helm_deploy/values-dev.yaml
@@ -12,12 +12,18 @@ generic-service:
     PRISON_API_URL: https://prison-api-dev.prison.service.justice.gov.uk
     DOCUMENT_MANAGEMENT_API_URL: "https://document-api-dev.hmpps.service.justice.gov.uk"
     ADJUSTMENTS_API_URL: "https://adjustments-api-dev.hmpps.service.justice.gov.uk"
+    COURT-REGISTER_API_URL: https://court-register-api-dev.hmpps.service.justice.gov.uk
     INGRESS_URL: *ingressHost
     ENVIRONMENT_NAME: DEV
     SPRINGDOC_API-DOCS_ENABLED: "true"
     SPRINGDOC_SWAGGER-UI_ENABLED: "true"
     FEATURES_APPEALS_ENABLED: "true"
     FEATURES_MARK_IMMIGRATION_CASES_INACTIVE_ENABLED: "true"
+    HMPPS_SAR_ENABLED: "false"
+    HMPPS_SAR_MODE_ALL-SAR-DATA_ENABLED: "false"
+    HMPPS_SAR_ADDITIONALACCESSROLE: ROLE_VIEW_PRISONER_DATA
+    HMPPS_SAR_MODE_TEMPLATE.PATH: /templates/template_remand-and-sentencing-api.mustache
+    HMPPS_SAR_MODE_TEMPLATE.ENABLED: "false"
 
   resources:
     requests:
