+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/config/SubjectAccessRequestConfiguration.kt
@@ -3,11 +3,9 @@ package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.config
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.CourtCaseUnsyncedSarRepository
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.ImmigrationDetentionUnsyncedSarRepository
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.alldata.CourtCaseSarRepository
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.alldata.ImmigrationDetentionSarRepository
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.alldata.RecallSarRepository
+import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.CourtCaseSarRepository
+import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.ImmigrationDetentionSarRepository
+import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.RecallSarRepository
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.CourtRegisterService
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.PersonService
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.service.subjectaccessrequest.AllPrisonerDetailsService
@@ -47,7 +45,7 @@ class SubjectAccessRequestConfiguration {
   )
   @Bean
   fun unsyncedPrisonerDetailsService(
-    immigrationDetentionUnsyncedSarRepository: ImmigrationDetentionUnsyncedSarRepository,
-    courtCaseUnsyncedSarRepository: CourtCaseUnsyncedSarRepository,
+    immigrationDetentionUnsyncedSarRepository: ImmigrationDetentionSarRepository,
+    courtCaseUnsyncedSarRepository: CourtCaseSarRepository,
   ): PrisonerDetailsService = UnsyncedPrisonerDetailsService(immigrationDetentionUnsyncedSarRepository, courtCaseUnsyncedSarRepository)
 }
