+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/jpa/entity/subjectaccessrequest/NextCourtAppearanceSarEntity.kt
@@ -7,20 +7,22 @@ import org.hibernate.annotations.Immutable
 import org.hibernate.annotations.Subselect
 import org.hibernate.annotations.Synchronize
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.config.ConditionalOnSarEnabled
+import java.time.LocalDate
 
 @ConditionalOnSarEnabled
-@Entity
 @Immutable
+@Entity
 @Subselect(
   """
-  select id, prisoner_id 
-  from court_case 
-  where status_id not in ('DELETED', 'DUPLICATE')""",
+  select id
+  ,appearance_date
+  from next_court_appearance""",
 )
-@Synchronize("court_case")
-class CourtCaseUnsyncedSarEntity(
+@Synchronize("next_court_appearance")
+class NextCourtAppearanceSarEntity(
   @Id
   @Column
-  var id: Int = 0,
-  var prisonerId: String,
+  val id: Int,
+  @Column
+  var appearanceDate: LocalDate,
 )
