+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/jpa/entity/subjectaccessrequest/RecallTypeSarEntity.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.subjectaccessrequest.alldata
+package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.subjectaccessrequest
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
@@ -14,17 +14,13 @@ import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.config.Condition
 @Subselect(
   """
   select id
-  ,outcome_type
-  ,outcome_name
-  from charge_outcome""",
+   ,code
+  from recall_type""",
 )
-@Synchronize("charge_outcome")
-class ChargeOutcomeSarEntity(
+@Synchronize("recall_type")
+class RecallTypeSarEntity(
   @Id
   @Column
-  var id: Int = 0,
-  @Column
-  var outcomeType: String,
-  @Column
-  var outcomeName: String,
+  var id: Int,
+  var code: String,
 )
