+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/jpa/entity/subjectaccessrequest/SentenceSarEntity.kt
@@ -1,53 +1,54 @@
-package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.subjectaccessrequest.alldata
+package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.subjectaccessrequest
 
-import jakarta.persistence.CascadeType
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
-import jakarta.persistence.FetchType
 import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.OneToMany
+import jakarta.persistence.OneToOne
 import org.hibernate.annotations.Immutable
 import org.hibernate.annotations.Subselect
 import org.hibernate.annotations.Synchronize
 import org.hibernate.proxy.HibernateProxy
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.config.ConditionalOnSarEnabled
-import java.time.ZonedDateTime
 
 @ConditionalOnSarEnabled
-@Entity
 @Immutable
+@Entity
 @Subselect(
   """
   select id
-   ,prisoner_id
-   ,latest_court_appearance_id
-   ,case_unique_identifier
+   ,charge_id
+   ,sentence_type_id
+   ,sentence_serve_type
    ,status_id
-   ,created_at
-   ,updated_at
-  from court_case
-  where status_id not in ('DELETED', 'DUPLICATE')
-  """,
+  from sentence""",
 )
-@Synchronize("court_case")
-class CourtCaseSarEntity(
+@Synchronize("sentence")
+class SentenceSarEntity(
   @Id
   @Column
-  var id: Int = 0,
-  var prisonerId: String,
-  @OneToMany(mappedBy = "courtCase", cascade = [CascadeType.ALL], orphanRemoval = true)
-  var appearances: MutableSet<CourtAppearanceSarEntity> = mutableSetOf(),
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "latest_court_appearance_id", referencedColumnName = "id", nullable = true)
-  var latestCourtAppearance: CourtAppearanceSarEntity?,
-  var caseUniqueIdentifier: String,
+  var id: Int,
+  @Suppress("JpaDataSourceORMInspection")
+  @ManyToOne
+  @JoinColumn(name = "charge_id")
+  var charge: ChargeSarEntity,
+  @Suppress("JpaDataSourceORMInspection")
+  @OneToOne
+  @JoinColumn(name = "sentence_type_id")
+  var sentenceType: SentenceTypeSarEntity?,
+  @Column
+  var sentenceServeType: String,
+  @Column
   var statusId: String,
-  var createdAt: ZonedDateTime,
-  var updatedAt: ZonedDateTime,
+  @Suppress("JpaDataSourceORMInspection")
+  @OneToMany
+  @JoinColumn(name = "sentence_id")
+  var periodLengths: MutableSet<PeriodLengthSarEntity> = mutableSetOf(),
+  @OneToMany(mappedBy = "sentence")
+  var recallSentences: MutableSet<RecallSentenceSarEntity> = mutableSetOf(),
 ) {
-
   final override fun equals(other: Any?): Boolean {
     if (this === other) return true
     if (other == null) return false
@@ -56,7 +57,7 @@ class CourtCaseSarEntity(
     val thisEffectiveClass =
       if (this is HibernateProxy) this.hibernateLazyInitializer.persistentClass else this.javaClass
     if (thisEffectiveClass != oEffectiveClass) return false
-    other as CourtCaseSarEntity
+    other as SentenceSarEntity
 
     return id == other.id
   }
