--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/jpa/repository/subjectaccessrequest/ImmigrationDetentionUnsyncedSarRepository.kt
@@ -1,11 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest
-
-import org.springframework.data.repository.CrudRepository
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.config.ConditionalOnSarEnabled
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.subjectaccessrequest.ImmigrationDetentionUnsyncedSarEntity
-
-@ConditionalOnSarEnabled
-interface ImmigrationDetentionUnsyncedSarRepository : CrudRepository<ImmigrationDetentionUnsyncedSarEntity, String> {
-
-  fun findByPrisonerId(prisonerId: String): List<ImmigrationDetentionUnsyncedSarEntity>
-}
