+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/service/subjectaccessrequest/UnsyncedPrisonerDetailsServiceTests.kt
@@ -6,22 +6,22 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.controller.dto.subjectaccessrequest.ImmigrationDetention
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.controller.dto.subjectaccessrequest.Prisoner
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.subjectaccessrequest.ImmigrationDetentionUnsyncedSarEntity
+import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.subjectaccessrequest.ImmigrationDetentionSarEntity
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.enum.ImmigrationDetentionNoLongerOfInterestType
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.CourtCaseUnsyncedSarRepository
-import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.ImmigrationDetentionUnsyncedSarRepository
+import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.CourtCaseSarRepository
+import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.repository.subjectaccessrequest.ImmigrationDetentionSarRepository
 import java.time.LocalDate
 
 class UnsyncedPrisonerDetailsServiceTests {
 
-  private val immigrationDetentionUnsyncedRepository = mockk<ImmigrationDetentionUnsyncedSarRepository>()
-  private val courtCaseUnsyncedSarRepository = mockk<CourtCaseUnsyncedSarRepository>()
+  private val immigrationDetentionUnsyncedRepository = mockk<ImmigrationDetentionSarRepository>()
+  private val courtCaseUnsyncedSarRepository = mockk<CourtCaseSarRepository>()
 
   @Test
   fun `should return all Immigration Detention Prisoner Details`() {
     every { immigrationDetentionUnsyncedRepository.findByPrisonerId("44959") } returns listOf(
-      ImmigrationDetentionUnsyncedSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "asdasdas", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
-      ImmigrationDetentionUnsyncedSarEntity(456, "44959", LocalDate.of(2026, 12, 19), "asdasdas", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
+      ImmigrationDetentionSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "asdasdas", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
+      ImmigrationDetentionSarEntity(456, "44959", LocalDate.of(2026, 12, 19), "asdasdas", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
     )
     every { courtCaseUnsyncedSarRepository.existsByPrisonerId("44959") } returns true
     val service = UnsyncedPrisonerDetailsService(immigrationDetentionUnsyncedRepository, courtCaseUnsyncedSarRepository)
@@ -41,8 +41,8 @@ class UnsyncedPrisonerDetailsServiceTests {
   @Test
   fun `should return all Immigration Detention Prisoner Details from date`() {
     every { immigrationDetentionUnsyncedRepository.findByPrisonerId("44959") } returns listOf(
-      ImmigrationDetentionUnsyncedSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "124222111", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
-      ImmigrationDetentionUnsyncedSarEntity(456, "44959", LocalDate.of(2026, 12, 19), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
+      ImmigrationDetentionSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "124222111", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
+      ImmigrationDetentionSarEntity(456, "44959", LocalDate.of(2026, 12, 19), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
     )
     every { courtCaseUnsyncedSarRepository.existsByPrisonerId("44959") } returns true
     val service = UnsyncedPrisonerDetailsService(immigrationDetentionUnsyncedRepository, courtCaseUnsyncedSarRepository)
@@ -61,8 +61,8 @@ class UnsyncedPrisonerDetailsServiceTests {
   @Test
   fun `should return all Immigration Detention Prisoner Details to date`() {
     every { immigrationDetentionUnsyncedRepository.findByPrisonerId("44959") } returns listOf(
-      ImmigrationDetentionUnsyncedSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "124222111", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
-      ImmigrationDetentionUnsyncedSarEntity(456, "44959", LocalDate.of(2026, 12, 19), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
+      ImmigrationDetentionSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "124222111", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
+      ImmigrationDetentionSarEntity(456, "44959", LocalDate.of(2026, 12, 19), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
     )
     every { courtCaseUnsyncedSarRepository.existsByPrisonerId("44959") } returns true
     val service = UnsyncedPrisonerDetailsService(immigrationDetentionUnsyncedRepository, courtCaseUnsyncedSarRepository)
@@ -81,9 +81,9 @@ class UnsyncedPrisonerDetailsServiceTests {
   @Test
   fun `should return all Immigration Detention Prisoner Details from date to date`() {
     every { immigrationDetentionUnsyncedRepository.findByPrisonerId("44959") } returns listOf(
-      ImmigrationDetentionUnsyncedSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "124222111", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
-      ImmigrationDetentionUnsyncedSarEntity(456, "44959", LocalDate.of(2026, 11, 10), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
-      ImmigrationDetentionUnsyncedSarEntity(456, "44959", LocalDate.of(2026, 12, 19), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "dolor sit amet"),
+      ImmigrationDetentionSarEntity(34, "44959", LocalDate.of(2020, 1, 1), "124222111", ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), ""),
+      ImmigrationDetentionSarEntity(456, "44959", LocalDate.of(2026, 11, 10), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "Lorem ipsum"),
+      ImmigrationDetentionSarEntity(456, "44959", LocalDate.of(2026, 12, 19), null, ImmigrationDetentionNoLongerOfInterestType.BRITISH_CITIZEN.toString(), "dolor sit amet"),
     )
     every { courtCaseUnsyncedSarRepository.existsByPrisonerId("44959") } returns true
     val service = UnsyncedPrisonerDetailsService(immigrationDetentionUnsyncedRepository, courtCaseUnsyncedSarRepository)
